Reorganize tests
Travis CI:
- Introduce test suites for any PHP version as well as for PHP >= 7.0.0
- Simplified setting up PHP

AppVeyor:
- Do not use all known PHP versions on Windows' build matrix.
  It is irrelevant
