+++ appveyor.yml
@@ -2,31 +2,6 @@ version: 0.11.7-{build}
 
 environment:
     matrix:
-        - PHP_MINOR: 7.0
-          VC_VERSION: 14
-          BUILD_TYPE: Win32
-          APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2013
-
-        - PHP_MINOR: 7.0
-          VC_VERSION: 14
-          BUILD_TYPE: nts-Win32
-          APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2013
-
-        - PHP_MINOR: 7.1
-          VC_VERSION: 14
-          BUILD_TYPE: Win32
-          APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2013
-
-        - PHP_MINOR: 7.1
-          VC_VERSION: 14
-          BUILD_TYPE: nts-Win32
-          APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2013
-
-        - PHP_MINOR: 7.2
-          VC_VERSION: 15
-          BUILD_TYPE: Win32
-          APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2015
-
         - PHP_MINOR: 7.2
           VC_VERSION: 15
           BUILD_TYPE: nts-Win32
